Add constraint-compiler
Compile emitters and verifier as our internal constraint rule representation. It is a key-value pair of <emitter, [emitters list, verifier]>. The emitter list tracks all emitters to ensure the verifier receives sufficient values from the emitter. Every time an emitter emits values, it is removed from the list (through constraint-updater); when the list is empty, the verifier has all values and is ready to verify the constraint (through constraint-checker). We use the predicate (goal function) name and 0 or 1 as the emitter name. 0 or 1 depends on the emitter coming from a positive(0) or negative (1) goal. We can only keep a unique emitter name as the key to save space. To speed up the remove operation on the list, we put the emitter used as key at the first.