Add constraint-checker
The constraint is ready to check when the verifier receives the values from the last emitter. The verifier accumulates the previous values, which are stored in L. If the verifier only needs one emitter, it stores in constraint-rules.

Therefore, we first combine constraint rules with L to filter out those verifiers that are ready to update after receiving the final values from the emitter. We use constraint-constructor to complete the verifier's continuation. Then, the verifier is ready to use eval to get the constraint checking outcome.