Fine grind global checking (Part 2)
Any goal functions (emitters) touched by the `constainto` are treated as global constraints; hence, we add these emitters to global-checking rules.

In the future (Part 3/4), I plan to modify the `run` interface so it can pick a subset of global-checking rules to check. As a result, we allow multiple programs to be loaded into the same environment. Also, the `run` interface can use a dependency graph analysis to determine whether the `noto` is stratified.