Add self-aware program analysis twins (body)
Syntax analysis to find stratified negations (noto + loop) in the program. This is the first part, body translation, where the body of each clause is rewritten. The head definition happens with `defineo`, where we assign different semantics to the loops. In the body translation, we remove the variables and append a suffix (+/-) to the goal's name to create twins.  Refer to the stratified negation definition for the justification. A stratified negation means the negation (noto) and the loop in the program do not mix. There are two levels of stratification: compilation time and execution time. Compilation time means the stratification can be identified by analyzing the input program with little effort, through simulation.  We identify the compilation time here by removing the variables in the program and creating two twins.