Add self-aware program analysis twins (head)
Syntax analysis to find stratified negations (noto + loop) in the program. This is the second part, where the head definition of each clause is provided. The body translation happens with `positive/negative-twin`, where we remove the variables and append a suffix (+/-) to the goal's name to create twins. In the head definition, we assign different semantics to the loops. Refer to the stratified negation definition for the justification. A stratified negation means the negation (noto) and the loop in the program do not mix. There are two levels of stratification: compilation time and execution time. Compilation time refers to the stratification that can be identified by analyzing the input program with minimal effort through simulation.  We identify the compilation time here by dividing the twins into different semantics. The positive twin returns true on a negative loop, but the negative twin returns false. Therefore, the twins produce the same results for stratified programs and produce different results for normal programs. [ToDo] This works at the syntax level only; add runtime level analysis later.