Add a knight's tour example.
This is a classic search and backtracking example, and a special case of the Hamiltonian path. It has a linear time solution if we use a proper heuristic algorithm.

The first solution, written in constraint logic programming (CLP), cannot produce answers for a 5*5 board.

The second solution, written in answer set programming (ASP), uses conflict-driven nogood learning (CDNL) to produce 1728 answers in 52 seconds.

The third solution, written in stableKanren without a heuristic, produces one answer in 1000 seconds.

We use this example as a starting point to apply heuristic algorithms to stableKanren.