Add constant constraint support (Part 5)
Add constant constraint support as mentioned in #13

We put everything together. In `constraint-updater` and `constraint-checker`, we use `constraint-emitter-filter` to filter out matching variable and constant emitters. Then in `constraint-propagator`, we use `constraint-emitter-reorder` to move the proper emitter to the first of the emitter list, `constraint-emitter-remove-constants` to remove any constants in the parameter list and the corresponding values.